proyecto_arqui/mqtt_suscriber/auctions/requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
typing
datetime
paho-mqtt
psycopg2-binary
requests
